backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "engineering-oracle"
version = "0.1.0"
description = "Multi-Agent RAG Research & Benchmark Engine"
requires-python = ">=3.11"
dependencies = [
    # Framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sse-starlette>=2.0.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    # LangGraph / LangChain
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-community>=0.3.0",
    "langchain-text-splitters>=0.3.0",
    # Search
    "duckduckgo-search>=6.0.0",
    "tavily-python>=0.5.0",
    # Scraping
    "scrapy>=2.11",
    "scrapy-playwright>=0.0.40",
    "beautifulsoup4>=4.12",
    "httpx>=0.27",
    # Git
    "gitpython>=3.1",
    # PDF
    "pymupdf4llm>=0.0.10",
    "pymupdf>=1.24",
    # Vector DB & Embeddings
    "chromadb>=0.5.0",
    "sentence-transformers>=3.0",
    # Database
    "sqlalchemy>=2.0",
    "aiosqlite>=0.20",
    # Utils
    "tenacity>=8.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-httpx>=0.30",
    "ruff>=0.5",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
